python-simulator-over-mqtt-master/requirements.txt
paho-mqtt==1.6.1
pandas==2.2.2
openpyxl==3.1.5
XlsxWriter==3.2.0
